Feature: add support for non-integers

+++ src/main/java/com/lampirg/calculator/logic/expression/BinaryExpression.java
@@ -3,15 +3,14 @@ package com.lampirg.calculator.logic.expression;
 import lombok.RequiredArgsConstructor;
 
 import java.util.function.BiFunction;
-import java.util.function.Function;
 
 @RequiredArgsConstructor
 public class BinaryExpression {
-    private final int x1;
-    private final int x2;
-    protected final BiFunction<Integer, Integer, Integer> operation;
+    private final double x1;
+    private final double x2;
+    protected final BiFunction<Double, Double, Double> operation;
 
-    public int compute() {
+    public double compute() {
         return operation.apply(x1, x2);
     }
 }
